+++ src/main/java/com/lotte/carts/controller/CartController.java
@@ -1,6 +1,6 @@
 package com.lotte.carts.controller;
 
-import com.lotte.carts.dto.response.CartResponseDto;
+import com.lotte.carts.dto.response.CartResponse;
 import com.lotte.carts.service.CartService;
 import java.time.LocalDateTime;
 import org.slf4j.Logger;
@@ -8,6 +8,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -21,8 +22,9 @@ public class CartController {
     }
 
     @GetMapping("/carts/{cartNo}")
-    public ResponseEntity<CartResponseDto> getCart(@PathVariable Integer cartNo) {
+    public ResponseEntity<CartResponse.CartItemsDto> getCart(@PathVariable Integer cartNo) {
         logger.info("CartController.getCart(cartNo): {}", LocalDateTime.now());
         return ResponseEntity.ok(cartService.getCart(cartNo));
     }
+
 }
