+++ src/main/resources/sqls/Cart.xml
@@ -7,7 +7,7 @@
 
     <!-- 장바구니 목록 (장바구니 아이템 목록) 조회 쿼리 -->
     <select id="selectCartItemsByCartNo"
-      resultType="com.lotte.carts.dto.response.CartResponseDto$CartDto"
+      resultType="com.lotte.carts.dto.CartItemDto"
       parameterType="Integer">
 
         SELECT cart_item_no cartItemNo, p.product_no productNo, product_name productName,
